Move FlipperComposeView to nfceditor module
