+++ components/nfceditor/impl/build.gradle.kts
@@ -11,6 +11,7 @@ dependencies {
     implementation(projects.components.core.ui.lifecycle)
     implementation(projects.components.core.ui.theme)
     implementation(projects.components.core.ui.res)
+    implementation(projects.components.core.navigation)
     implementation(projects.components.core.log)
     implementation(projects.components.core.di)
