+++ components/nfceditor/impl/src/main/java/com/flipperdevices/nfceditor/impl/fragments/NfcEditorFragment.kt
@@ -1,5 +1,9 @@
 package com.flipperdevices.nfceditor.impl.fragments
 
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
@@ -10,6 +14,18 @@ import com.flipperdevices.core.ui.fragment.ComposeFragment
 import com.flipperdevices.nfceditor.impl.composable.ComposableNfcEditor
 
 class NfcEditorFragment : ComposeFragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return FlipperComposeView(requireContext()).apply {
+            setContent {
+                ComposeViewRenderWithTheme()
+            }
+        }
+    }
+
     @Composable
     override fun RenderView() {
         Box(
